pyproject.toml
[project]
name = "voice-cv-agent"
version = "0.1.0"
description = "Voice agent to ask question about my resume and more."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "groq>=0.37.1",
    "httpx>=0.28.1",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.18",
    "langchain-groq>=1.1.2",
    "pymupdf>=1.27.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "twilio>=9.10.3",
    "uvicorn>=0.41.0",
]
